Coins!!!!!!!!
Coins will now be randomly generated in value based on enemyLevel * 3
Coins carry over to Eterius, save and load, and reset. and used
Coins need SPECIFIC PREFAB BASED ON THE AMOUNT GENERATED

+++ PurrfectForgeGameStudios1/Assets/Scripts/Combat/ScriptableScripts/ScriptableEnemies.cs
@@ -12,7 +12,7 @@ public class ScriptableEnemies : ScriptableObject
     public enum Type { Normal, Ranged, Boss, Wave}
     public Type type;
     public string EnemyName;
-    public float Level;
+    public int Level;
     public float HP;
     public float Defense;
     public float LineOfSight;
